Upgrade to React Router v4 (#283)
* initial setup of rr4

* :star2: feat(global): wip converting to react-router 4

beginning the process of converting to RR4. For the most part, the majority of the routing is
complete. Data fetching isn't fully there in the dashboard. Theres a :bear: of a problem with the
dashboard routes not rendering as children. the sidebar goes away and the whole page re-renders on
route change.

* :star2: feat(global): add back more missing datch fetching

* progress toward resolving the RR upgrade issues

* filemanager data and attachments actions

* :star2: feat(global): All routes and data fetching working correctly for RR4 upgra

282

* :star2: feat(routing): continuing implementation of RR4

#282

* nearly finished with upgrade to RR4  #282

* remove comments. will use disqus

* :star2: feat(global): fixes for RR4 implementation. only aspect left is auth check

#282

* :star2: feat(global): React Router 4 upgrade finished

#282